Add lexer
This creates a new lexer file for the miniml language, which includes regular expressions and token definitions, as well as a lexer function place holder.

Additionally, this commit modifies the dune file to include the batteries library and also changes the main program to call the new lexer function with a test input.