Move Lexer code from bin to lib, update dependencies and add tests
The Lexer for the miniml language was previously in bin/lexer.ml. It was moved to lib/lexer.ml. The dependencies have been updated in the dune files for bin and lib, as well as in the main.ml file. Tests for the lexer were added in test/test_lexer.ml. The dune file for tests was also updated.