Integrate Parser into build process and rename 'lib' to 'Lexical'
This commit integrates the parser into the build process by adding calls to the parser function in the main function. It also changes the 'lexer' function in lexer.ml to return a list of tokens instead of void and improves error handling.

Separately, there's been renaming of the 'lib' directory and its files to 'Lexical' to better indicate its function in the codebase. This is reflected in the 'lib' directory's 'dune' file as well.

Additionally, a parser file has been added with initial parsing functionality, which simply prints the input tokens.